Introduce vAccelrt library in Unikraft
vAccel is a runtime system for hardware acceleration. It provides an API with
a set of functions that the runtime is able to offload to hardware
acceleration devices.

In case of Unikraft vAccel will be used as a transport layer between the
guest and the host, where the request will be offloaded in the relative
hardware.

As a result, in the case of Unikraft only the virtio plugin of vAccelrt is
used.

Signed-off-by: Charalampos Mainas <charalampos.mainas@gmail.com>