Introduce sgemm - blas operation
There is an incompatibility in the number of arguments between the
vaccel_sgemm function in vaccelrt and virtio_sgemm in virtio plugin.
The difference is two arguments, alpha and beta, which were introduced
in vaccelrt.

For the time being, we pass as many arguments as in the vaccelrt
version. We might need to revisit this in the future.

Signed-off-by: Charalampos Mainas <cmainas@nubificus.co.uk>